app/pyproject.toml
[project]
name = "sullivan-backend"
version = "0.1.0"
description = "& Sullivan - AI Healthcare Backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0,<1.0",
    "uvicorn[standard]>=0.30.0,<1.0",
    "tortoise-orm>=0.25.0,<2.0",
    "PyJWT>=2.10.0,<3.0",
    "cryptography>=44.0.0",
    "bcrypt>=4.2.0,<5.0",
    "python-multipart>=0.0.20",
    "pydantic>=2.9.0,<3.0",
    "pydantic-settings>=2.7.0,<3.0",
    "httpx>=0.27.0,<1.0",
    "openai>=1.60.0,<2.0",
    "email-validator>=2.3.0,<3.0",
    "arq>=0.26.0,<1.0",
    "redis>=5.0.0,<8.0",
    "aiomysql>=0.2.0,<1.0",
    "python-dotenv>=1.0.0,<2.0",
    "slowapi>=0.1.9",
    "faiss-cpu>=1.9.0,<2.0",
    "numpy>=1.26.0,<3.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--ignore=tests/integration"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.27.0",
    "ruff>=0.9.0",
    "aiosqlite>=0.21.0",
]